requirements.txt
Django==1.11.9
djangorestframework==3.7.7
model-mommy==1.5.0
django-fixture-magic==0.1.3
